# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pragma-api",
    "description": "0D Master API for vaults",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://{environment}.0d.finance/api/v1",
      "variables": {
        "environment": {
          "default": "devnet",
          "enum": [
            "devnet",
            "production"
          ]
        }
      }
    }
  ],
  "paths": {
...
    "/vaults/{vault_id}/composition": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_composition",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group_by",
            "in": "query",
            "description": "Aggregate by platform or asset",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "platform"
          }
        ],
        "responses": {
          "200": {
            "description": "Current vault composition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultCompositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/composition/series": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_composition_series",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Time period for composition series",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "30d"
          },
          {
            "name": "group_by",
            "in": "query",
            "description": "Group by platform or asset",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "platform"
          }
        ],
        "responses": {
          "200": {
            "description": "Historical composition data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultCompositionSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/kpis": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_kpis",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Time period for KPI calculation",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "all"
          }
        ],
        "responses": {
          "200": {
            "description": "Vault performance KPIs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultKpisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompositionPosition": {
        "type": "object",
        "required": [
          "platform",
          "asset",
          "symbol",
          "pct",
          "apy_est_pct"
        ],
        "properties": {
          "apy_est_pct": {
            "type": "number",
            "format": "double"
          },
          "asset": {
            "type": "string"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "pct": {
            "type": "number",
            "format": "double"
          },
          "platform": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          }
        }
      },
      "CompositionSeriesPoint": {
        "type": "object",
        "required": [
          "t",
          "weights_pct"
        ],
        "properties": {
          "t": {
            "type": "string"
          },
          "weights_pct": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
...
      "VaultCompositionResponse": {
        "type": "object",
        "required": [
          "as_of",
          "positions"
        ],
        "properties": {
          "as_of": {
            "type": "string"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompositionPosition"
            }
          }
        }
      },
      "VaultCompositionSeriesResponse": {
        "type": "object",
        "required": [
          "timeframe",
          "group_by",
          "labels",
          "points"
        ],
        "properties": {
          "group_by": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompositionSeriesPoint"
            }
          },
          "timeframe": {
            "type": "string"
          }
        }
      },
      "VaultKpisResponse": {
        "type": "object",
        "required": [
          "cumulative_pnl_usd",
          "max_drawdown_pct",
          "sharpe",
          "profit_share_bps"
        ],
        "properties": {
          "cumulative_pnl_usd": {
            "type": "string"
          },
          "max_drawdown_pct": {
            "type": "number",
            "format": "double"
          },
          "profit_share_bps": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "sharpe": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  },
...
}